+++ src/Extension.php
@@ -10,8 +10,9 @@
 
 namespace Pronamic\WordPress\Pay\Extensions\RestrictContentPro;
 
+use Pronamic\WordPress\DateTime\DateTime;
 use Pronamic\WordPress\Pay\Core\Recurring;
-use Pronamic\WordPress\Pay\Core\Statuses;
+use Pronamic\WordPress\Pay\Payments\PaymentStatus as Core_PaymentStatus;
 use Pronamic\WordPress\Pay\Payments\Payment;
 use RCP_Member;
 use RCP_Payments;
@@ -21,7 +22,7 @@ use WP_Query;
  * Extension
  *
  * @author  Reüel van der Steege
- * @version 2.1.1
+ * @version 2.1.5
  * @since   1.0.0
  */
 class Extension {
@@ -175,7 +176,7 @@ class Extension {
 			return $url;
 		}
 
-		if ( Statuses::SUCCESS !== $payment->get_status() ) {
+		if ( Core_PaymentStatus::SUCCESS !== $payment->get_status() ) {
 			return $url;
 		}
 
@@ -216,23 +217,13 @@ class Extension {
 		);
 
 		switch ( $core_status ) {
-			case Statuses::CANCELLED:
-				$rcp_payments->update( $rcp_payment_id, $rcp_payment_data );
-
-				break;
-			case Statuses::EXPIRED:
-				$rcp_payments->update( $rcp_payment_id, $rcp_payment_data );
-
-				break;
-			case Statuses::FAILURE:
-				$rcp_payments->update( $rcp_payment_id, $rcp_payment_data );
-
-				break;
-			case Statuses::SUCCESS:
+			case Core_PaymentStatus::CANCELLED:
+			case Core_PaymentStatus::EXPIRED:
+			case Core_PaymentStatus::FAILURE:
 				$rcp_payments->update( $rcp_payment_id, $rcp_payment_data );
 
 				/**
-				 * Find and renew the Restrict Content Pro membership.
+				 * Find and cancel the Restrict Content Pro membership.
 				 *
 				 * @link https://gitlab.com/pronamic-plugins/restrict-content-pro/blob/3.0.10/includes/memberships/membership-functions.php#L15-29
 				 * @link https://gitlab.com/pronamic-plugins/restrict-content-pro/blob/3.0.10/includes/class-rcp-payments.php#L75
@@ -243,15 +234,19 @@ class Extension {
 				$rcp_membership = rcp_get_membership( $rcp_payment->membership_id );
 
 				if ( false !== $rcp_membership ) {
-					$subscription = $payment->get_subscription();
+					// Set expiration date to yesterday.
+					$rcp_membership->expire();
 
-					$recurring = empty( $subscription ) ? false : true;
-
-					$rcp_membership->renew( $recurring );
+					// Set status to `pending` to prevent access to restricted content.
+					$rcp_membership->set_status( 'pending' );
 				}
 
 				break;
-			case Statuses::OPEN:
+			case Core_PaymentStatus::SUCCESS:
+				$rcp_payments->update( $rcp_payment_id, $rcp_payment_data );
+
+				break;
+			case Core_PaymentStatus::OPEN:
 				// Nothing to do?
 				break;
 		}
@@ -293,7 +288,9 @@ class Extension {
 			return;
 		}
 
-		$core_status = MembershipStatus::to_core( $new_status );
+		$core_status = MembershipStatus::to_core_subscription_status( $new_status );
+
+		$note = null;
 
 		switch ( $new_status ) {
 			case MembershipStatus::ACTIVE:
@@ -339,14 +336,17 @@ class Extension {
 		while ( $query->have_posts() ) {
 			$query->the_post();
 
+			// Get subscription.
 			$subscription = get_pronamic_subscription( get_the_ID() );
 
 			if ( null === $subscription ) {
 				continue;
 			}
 
+			// Set subscription status.
 			$subscription->set_status( $core_status );
 
+			// Add note.
 			$subscription->add_note( $note );
 
 			$subscription->save();
@@ -360,9 +360,10 @@ class Extension {
 	 *
 	 * @link https://gitlab.com/pronamic-plugins/restrict-content-pro/blob/3.0.10/includes/memberships/class-rcp-membership.php#L2239-2248
 	 *
-	 * @param bool           $can_cancel    Whether or not this membership can be cancelled.
-	 * @param int            $membership_id ID of the membership.
-	 * @param RCP_Membership $membership    Membership object.
+	 * @param bool            $can_cancel    Whether or not this membership can be cancelled.
+	 * @param int             $membership_id ID of the membership.
+	 * @param \RCP_Membership $membership    Membership object.
+	 *
 	 * @return bool
 	 */
 	public function rcp_membership_can_cancel( $can_cancel, $membership_id, $membership ) {
@@ -527,8 +528,8 @@ class Extension {
 	 * Get Restrict Content Pro mebership from payment.
 	 *
 	 * @param Payment $payment Pronamic Pay payment.
-	 * @return RCP_Membership|null
-	 * @throws Exception When Restrict Content Pro membership can not be found.
+	 * @return \RCP_Membership|null
+	 * @throws \Exception When Restrict Content Pro membership can not be found.
 	 */
 	private function get_rcp_membership_from_payment( Payment $payment ) {
 		/**
@@ -545,11 +546,11 @@ class Extension {
 			 *
 			 * @link https://gitlab.com/pronamic-plugins/restrict-content-pro/blob/3.1/includes/customers/customer-functions.php#L15-34
 			 */
-			$rcp_customer = rcp_get_customer_by_user_id( $user_id );
+			$rcp_customer = \rcp_get_customer_by_user_id( $user_id );
 
 			if ( false === $rcp_customer ) {
-				throw new Exception(
-					sprintf(
+				throw new \Exception(
+					\sprintf(
 						'Could not find Restrict Content Pro customer for user ID: %s.',
 						$user_id
 					)
@@ -561,11 +562,11 @@ class Extension {
 			 *
 			 * @link https://gitlab.com/pronamic-plugins/restrict-content-pro/blob/3.1/includes/customers/customer-functions.php#L280-320
 			 */
-			$rcp_membership = rcp_get_customer_single_membership( $rcp_customer->get_id() );
+			$rcp_membership = \rcp_get_customer_single_membership( $rcp_customer->get_id() );
 
 			if ( false === $rcp_membership ) {
-				throw new Exception(
-					sprintf(
+				throw new \Exception(
+					\sprintf(
 						'Could not find Restrict Content Pro membership for customer ID: %s.',
 						$rcp_customer->get_id()
 					)
@@ -588,11 +589,11 @@ class Extension {
 			 *
 			 * @link https://gitlab.com/pronamic-plugins/restrict-content-pro/blob/3.0.10/includes/memberships/membership-functions.php#L15-29
 			 */
-			$rcp_membership = rcp_get_membership( $membership_id );
+			$rcp_membership = \rcp_get_membership( $membership_id );
 
 			if ( false === $rcp_membership ) {
-				throw new Exception(
-					sprintf(
+				throw new \Exception(
+					\sprintf(
 						'Could not find Restrict Content Pro membership with ID: %s.',
 						$membership_id
 					)
@@ -614,7 +615,7 @@ class Extension {
 	 * @link https://github.com/wp-pay/core/blob/2.1.6/src/Payments/PaymentsDataStoreCPT.php#L234
 	 *
 	 * @param Payment $payment Payment.
-	 * @throws Exception When Restrict Content Pro returns unexpected value.
+	 * @throws \Exception When Restrict Content Pro returns unexpected value.
 	 */
 	public function new_payment( Payment $payment ) {
 		$rcp_membership = $this->get_rcp_membership_from_payment( $payment );
@@ -628,7 +629,7 @@ class Extension {
 		 *
 		 * @link https://gitlab.com/pronamic-plugins/restrict-content-pro/blob/3.0.10/includes/class-rcp-payments.php#L55-191
 		 */
-		$rcp_payments = new RCP_Payments();
+		$rcp_payments = new \RCP_Payments();
 
 		$result = $rcp_payments->insert(
 			array(
@@ -640,8 +641,8 @@ class Extension {
 				'membership_id'    => $rcp_membership->get_id(),
 				'amount'           => $payment->get_total_amount()->get_value(),
 				// Transaction ID can not be null therefor we use `strval` to cast `null` to an empty string.
-				'transaction_id'   => strval( $payment->get_transaction_id() ),
-				'subscription'     => rcp_get_subscription_name( $rcp_membership->get_object_id() ),
+				'transaction_id'   => \strval( $payment->get_transaction_id() ),
+				'subscription'     => \rcp_get_subscription_name( $rcp_membership->get_object_id() ),
 				'subscription_key' => $rcp_membership->get_subscription_key(),
 				'object_type'      => 'subscription',
 				'object_id'        => $rcp_membership->get_object_id(),
@@ -650,14 +651,26 @@ class Extension {
 		);
 
 		if ( false === $result ) {
-			throw new Exception(
-				sprintf(
+			throw new \Exception(
+				\sprintf(
 					'Could not create Restrict Content Pro payment for payment %s.',
 					$payment->get_id()
 				)
 			);
 		}
 
+		// Renew membership.
+		$expiration = '';
+
+		$end_date = $payment->get_end_date();
+
+		if ( null !== $end_date ) {
+			$expiration = $end_date->format( DateTime::MYSQL );
+		}
+
+		$rcp_membership->renew( true, 'active', $expiration );
+
+		// Set source.
 		$rcp_payment_id = $result;
 
 		$payment->source    = 'rcp_payment';
@@ -670,7 +683,7 @@ class Extension {
 	 * Update payment.
 	 *
 	 * @param Payment $payment Payment.
-	 * @throws Exception When Restrict Content Pro returns unexpected value.
+	 * @throws \Exception When Restrict Content Pro returns unexpected value.
 	 */
 	public function update_payment( Payment $payment ) {
 		/**
@@ -686,19 +699,19 @@ class Extension {
 		 *
 		 * @link https://gitlab.com/pronamic-plugins/restrict-content-pro/blob/master/includes/class-rcp-payments.php#L219-284
 		 */
-		$rcp_payments = new RCP_Payments();
+		$rcp_payments = new \RCP_Payments();
 
 		$result = $rcp_payments->update(
 			$payment->source_id,
 			array(
 				'status'         => PaymentStatus::from_core( $payment->get_status() ),
-				'transaction_id' => strval( $payment->get_transaction_id() ),
+				'transaction_id' => \strval( $payment->get_transaction_id() ),
 			)
 		);
 
 		if ( false === $result ) {
-			throw new Exception(
-				sprintf(
+			throw new \Exception(
+				\sprintf(
 					'Could not update Restrict Content Pro payment for payment %s.',
 					$payment->get_id()
 				)
@@ -714,7 +727,7 @@ class Extension {
 	 * @param RCP_Membership $membership Restrict Content Pro membership.
 	 */
 	public function rcp_edit_membership_after( $membership ) {
-		$query = new WP_Query(
+		$query = new \WP_Query(
 			array(
 				'post_type'     => 'pronamic_pay_subscr',
 				'post_status'   => 'any',
@@ -737,7 +750,7 @@ class Extension {
 
 		include __DIR__ . '/../views/edit-membership.php';
 
-		wp_reset_postdata();
+		\wp_reset_postdata();
 	}
 
 	/**
@@ -748,7 +761,7 @@ class Extension {
 	 * @param object $payment Restrict Content Pro payment.
 	 */
 	public function rcp_edit_payment_after( $payment ) {
-		$query = new WP_Query(
+		$query = new \WP_Query(
 			array(
 				'post_type'     => 'pronamic_payment',
 				'post_status'   => 'any',
@@ -771,6 +784,6 @@ class Extension {
 
 		include __DIR__ . '/../views/edit-payment.php';
 
-		wp_reset_postdata();
+		\wp_reset_postdata();
 	}
 }