+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a CHANGELOG](http://keepachangelog.com/).
 
+## [2.1.5] - 2019-10-07
+- Restrict Content Pro 3.0 is required.
+- Renew membership during `pronamic_pay_new_payment` routine and update membership expiration date and status on cancelled/expired/failed payment status update.
+
 ## [2.1.4] - 2019-08-26
 - Fixed support for Restrict Content Pro 3.0.
 - Updated packages.
@@ -36,7 +40,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2017-12-13
 - First release.
 
-[unreleased]: https://github.com/wp-pay-extensions/restrict-content-pro/compare/2.1.4...HEAD
+[unreleased]: https://github.com/wp-pay-extensions/restrict-content-pro/compare/2.1.5...HEAD
+[2.1.5]: https://github.com/wp-pay-extensions/restrict-content-pro/compare/2.1.4...2.1.5
 [2.1.4]: https://github.com/wp-pay-extensions/restrict-content-pro/compare/2.1.3...2.1.4
 [2.1.3]: https://github.com/wp-pay-extensions/restrict-content-pro/compare/2.1.2...2.1.3
 [2.1.2]: https://github.com/wp-pay-extensions/restrict-content-pro/compare/2.1.1...2.1.2